# lib/api-zod/src/generated/types/addLibraryMealPlanItemRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Keto Diet Doctor Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import type { AddLibraryMealPlanItemRequestMealType } from "./addLibraryMealPlanItemRequestMealType";

export interface AddLibraryMealPlanItemRequest {
  mealType: AddLibraryMealPlanItemRequestMealType;
  foodName: string;
  portionGrams: number;
  unit?: string;
  calories?: number;
  carbs?: number;
  fat?: number;
  protein?: number;
  notes?: string;
}