# lib/api-zod/src/generated/types/doctor.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Keto Diet Doctor Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface Doctor {
  id: number;
  username: string;
  name: string;
  email: string;
  specialty?: string;
}